operations-center: Move packages from cmd to internal
