+++ internal/cli/cmds/inventory/templates/cluster_names.gotmpl
