+++ internal/cli/cmds/inventory/templates/resource_tree.gotmpl
